# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Student Placement Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}

export interface Student {
  id: number;
  userId: number;
  name: string;
  email: string;
  rollNumber: string;
  department: string;
  cgpa: number;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  skills?: string;
  status: string;
  /** @nullable */
  tenthPercent?: number | null;
  /** @nullable */
  twelfthPercent?: number | null;
  /** @nullable */
  placedCompany?: string | null;
  /** @nullable */
  placedPackage?: number | null;
  createdAt: string;
}

export interface StudentInput {
  name: string;
  email: string;
  rollNumber: string;
  department: string;
  cgpa: number;
  phone?: string;
  skills?: string;
  tenthPercent?: number;
  twelfthPercent?: number;
}

export interface StudentUpdate {
  name?: string;
  phone?: string;
  cgpa?: number;
  skills?: string;
  resumeUrl?: string;
  tenthPercent?: number;
  twelfthPercent?: number;
  status?: string;
}

export interface StudentListResponse {
  students: Student[];
  total: number;
  page: number;
  limit: number;
}

export interface Company {
  id: number;
  name: string;
  industry: string;
  /** @nullable */
  website: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export interface CompanyInput {
  name: string;
  industry: string;
  website?: string;
  description?: string;
  logoUrl?: string;
  location?: string;
}

export interface CompanyUpdate {
  name?: string;
  industry?: string;
  website?: string;
  description?: string;
  logoUrl?: string;
  location?: string;
}

export interface CompanyListResponse {
  companies: Company[];
  total: number;
  page: number;
  limit: number;
}

export interface Job {
  id: number;
  title: string;
  companyId: number;
  companyName: string;
  /** @nullable */
  description?: string | null;
  minCgpa: number;
  /** @nullable */
  package?: number | null;
  /** @nullable */
  location?: string | null;
  status: string;
  driveDate: string;
  eligibleDepartments?: string;
  createdAt: string;
}

export interface JobInput {
  title: string;
  companyId: number;
  description?: string;
  minCgpa: number;
  package?: number;
  location?: string;
  status?: string;
  driveDate: string;
  eligibleDepartments?: string;
}

export interface JobUpdate {
  title?: string;
  description?: string;
  minCgpa?: number;
  package?: number;
  location?: string;
  status?: string;
  driveDate?: string;
  eligibleDepartments?: string;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
}

export interface Application {
  id: number;
  studentId: number;
  jobId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  companyName?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ApplicationInput {
  jobId: number;
}

export interface ApplicationStatusUpdate {
  status: string;
  notes?: string;
}

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  limit: number;
}

export interface Notification {
  id: number;
  title: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  targetRole?: string | null;
  createdAt: string;
}

export interface NotificationInput {
  title: string;
  message: string;
  targetRole?: string;
}

export interface MonthlyPlacement {
  month: string;
  count: number;
}

export interface DashboardStats {
  totalStudents: number;
  totalCompanies: number;
  totalJobs: number;
  totalApplications: number;
  selectedStudents: number;
  placementPercentage: number;
  avgPackage?: number;
  monthlyPlacements: MonthlyPlacement[];
}

export interface StudentDashboard {
  student: Student;
  applications: Application[];
  eligibleJobs: Job[];
  recentNotifications: Notification[];
}

export type ListStudentsParams = {
search?: string;
department?: string;
page?: number;
limit?: number;
};

export type ListCompaniesParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListJobsParams = {
search?: string;
status?: string;
page?: number;
limit?: number;
};

export type ListApplicationsParams = {
jobId?: number;
studentId?: number;
status?: string;
page?: number;
limit?: number;
};