# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Student Placement Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  industry: string;
  /** @nullable */
  website: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}