# lib/api-zod/src/generated/types/companyListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Student Placement Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';

export interface CompanyListResponse {
  companies: Company[];
  total: number;
  page: number;
  limit: number;
}